@@ -26,12 +26,17 @@
         }
     },
     "require": {
-        "php": ">=5.4",
-        "phpseclib/phpseclib": "^2.0",
+        "php": ">=7.4",
+        "phpseclib/phpseclib": "^3.0@stable",
         "comodojo/exceptions": "dev-master"
     },
     "require-dev": {
-        "phpunit/phpunit": "^4.0|^5.0",
-        "scrutinizer/ocular": "^1.0"
+        "phpunit/phpunit": "^9.0@stable"
+    },
+    "minimum-stability": "dev",
+    "scripts": {
+        "test": "sh tests/before.sh && vendor/bin/phpunit --coverage-text --coverage-clover=coverage.clover && sh tests/after.sh",
+        "test-multi": "/usr/bin/bash .docker/test.sh",
+        "scrutinizer": "wget https://scrutinizer-ci.com/ocular.phar && php ocular.phar code-coverage:upload --format=php-clover coverage.clover"
     }
 }
