@@ -1,4 +1,8 @@
-<?php namespace Comodojo\Cookies;
+<?php
+
+namespace Comodojo\Cookies;
+
+use \phpseclib3\Crypt\AES;
 
 /**
  * AES-encrypted cookie using client-specific key
@@ -18,35 +22,24 @@
  * THE SOFTWARE.
  */
 
-class SecureCookie extends EncryptedCookie {
+class SecureCookie extends EncryptedCookie
+{
 
     /**
      * Create a client-specific key using provided key,
      * the client remote address and (in case) the value of
      * HTTP_X_FORWARDED_FOR header
      *
-     * @param string $key
+     * {@inheritdoc}
      *
-     * @return string
      */
-    protected static function encryptKey($key) {
-
-        if ( isset($_SERVER['REMOTE_ADDR']) ) {
-
-            $client_hash = md5($_SERVER['REMOTE_ADDR'].(isset($_SERVER['HTTP_X_FORWARDED_FOR']) ? $_SERVER['HTTP_X_FORWARDED_FOR'] : ''), true);
-
-            $server_hash = md5($key, true);
-
-            $cookie_key = $client_hash.$server_hash;
-
-        } else {
-
-            $cookie_key = hash('sha256', $key);
-
-        }
-
-        return $cookie_key;
-
+    protected static function setupCipher(string $key): AES
+    {
+        $remoteAddress = $_SERVER['REMOTE_ADDR'] ?? '';
+        $xffHeader = $_SERVER['HTTP_X_FORWARDED_FOR'] ?? '';
+        $cipher = new AES('ecb');
+        $cipher->setKeyLength(256);
+        $cipher->setKey(hash('md5', $key.$remoteAddress.$xffHeader));
+        return $cipher;
     }
-
 }
